mining: add candidate lifecycle types, pure helpers, and capability traits
Move the mining-domain lifecycle vocabulary into crates/mining so the
service that will drive it never depends on node:

- coordinator.rs: GenerationKey, InFlight, CoordinatorState (bounded
  template cache + single-flight state), lifecycle constants, the
  MempoolSequenceWake trait, AppliedTipSource / MempoolSnapshotSource /
  ChainContextSource capability traits, a MiningService stub, and the
  pure projections (template_from_candidate, snapshot_for_selection,
  generation_race helpers, parse_long_poll_id, signet_info).
- node keeps proposal validation and solved-block submission unchanged;
  all seven BIP22/ApplyError projection helpers stay in
  crates/node/src/mining/submission.rs (ApplyError is node-owned).
- MempoolSequenceWake moves in one step: node imports the mining trait,
  GenerationKey drops from node's public exports, and zmq tests import
  the trait from bitcoin_rs_mining.
- node-side call sites (long-poll waits, candidate assembly, control
  surface, unit tests) switch to the moved items; no behavior change.

Verified: cargo check -p bitcoin-rs-node -p bitcoin-rs-rpc;
cargo test -p bitcoin-rs-mining; cargo test -p bitcoin-rs-node --lib
mining (25 passed).
