+++ crates/node/src/mining/candidate_template_tests.rs
@@ -1,8 +1,11 @@
 // CONTRACT: `docs/contracts/external-api.md#API-11` owns BIP22/BIP23
 // template capabilities, submitold, signet projection, and template rules.
 use alloc::sync::Arc;
+use bitcoin_rs_mining::CANDIDATE_CACHE_LIMIT;
 use bitcoin_rs_mining::Candidate;
+use bitcoin_rs_mining::CoordinatorState;
 use bitcoin_rs_mining::TemplateId;
+use bitcoin_rs_mining::template_from_candidate;
 use bitcoin_rs_primitives::Hash256;
 use bitcoin_rs_primitives::Network;
 use bitcoin_rs_primitives::Tx;
@@ -11,10 +14,11 @@ use bitcoin_rs_primitives::TxOut;
 #[test]
 fn candidate_cache_evicts_the_oldest_entry_at_the_bound() {
     use alloc::sync::Arc;
-    use bitcoin_rs_mining::Candidate;
+    use bitcoin_rs_mining::CANDIDATE_CACHE_LIMIT;
+use bitcoin_rs_mining::Candidate;
     use bitcoin_rs_primitives::{Amount, CompactTarget, LockTime, Script};
 
-    let mut state = super::CoordinatorState::new();
+    let mut state = CoordinatorState::new();
     let coinbase = Tx {
         version: 2,
         lock_time: LockTime::from_consensus(0),
@@ -25,7 +29,7 @@ fn candidate_cache_evicts_the_oldest_entry_at_the_bound() {
         }],
     };
     let mut first_id = None;
-    for seq in 0..=super::CANDIDATE_CACHE_LIMIT {
+    for seq in 0..=CANDIDATE_CACHE_LIMIT {
         let seq = u64::try_from(seq).unwrap_or(u64::MAX);
         let hash = Hash256::from_le_bytes(&[u8::try_from(seq).unwrap_or(0xff); 32]);
         let id = TemplateId::new(&hash, seq);
@@ -59,7 +63,7 @@ fn candidate_cache_evicts_the_oldest_entry_at_the_bound() {
         });
         state.cache_insert(id, candidate);
     }
-    assert_eq!(state.cache.len(), super::CANDIDATE_CACHE_LIMIT);
+    assert_eq!(state.cache.len(), CANDIDATE_CACHE_LIMIT);
     assert!(
         !state
             .cache
@@ -109,7 +113,7 @@ fn template_for(
     candidate: Candidate,
     submit_old: Option<bool>,
 ) -> bitcoin_rs_mining::BlockTemplate {
-    super::MiningCoordinator::template_from_candidate(
+    template_from_candidate(
         Network::Regtest,
         Arc::new(candidate),
         submit_old,
@@ -162,7 +166,7 @@ fn template_facts_follow_mutated_candidate_generation() {
 fn signet_template_carries_challenge_and_mandatory_rule() {
     use bitcoin_rs_mining::MiningRule;
 
-    let template = super::MiningCoordinator::template_from_candidate(
+    let template = template_from_candidate(
         Network::Signet,
         Arc::new(sample_candidate(
             Hash256::from_le_bytes(&[0x44; 32]),
